+++ src/Api/Repository.php
@@ -71,9 +71,16 @@ class Repository
     {
         $config = $this->configFor($apiName);
         $rootNamespace = $this->normalizeRootNamespace(array_get($config, 'namespace'));
-        $byResource = (bool) array_get($config, 'by-resource', true);
+        $byResource = array_get($config, 'by-resource', true);
+        $withType = true;
+
+        if ('false-0.x' === $byResource) {
+            $byResource = false;
+            $withType = false;
+        }
+
         $resources = (array) array_get($config, 'resources');
-        $resolver = $this->factory->createResolver($rootNamespace, $resources, $byResource);
+        $resolver = $this->factory->createResolver($rootNamespace, $resources, (bool) $byResource, $withType);
 
         $api = new Api(
             $this->factory,